requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
boto3==1.13.6
botocore==1.16.6
dj-database-url==0.5.0
Django==3.0.6
django-autoslug==1.9.7
django-cors-headers==3.2.1
django-filter==2.2.0
django-graphql-auth==0.3.10
django-graphql-jwt==0.3.0
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.15.2
graphene==2.1.8
graphene-django==2.9.1
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
jmespath==0.9.5
Pillow==7.1.2
promise==2.3
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
